LayoutTests/accessibility/mac/attributed-string/attributed-string-has-completion-annotation-expected.txt

This tests that custom annotations are applied to the accessibility string

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attrString.indexOf('AXTextCompletion = 1') is -1
PASS attrString.indexOf('AXTextCompletion = 1') != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
